Project2
Triangles are dope.
Easily the best shape.
I mean a lot of people like the fancy stuff,
like dodecahedron.
Or party tricks like the mobius,
but what are they without the humble triangle?
NOTHING.
Some folks would extend this argument and claim that by similar logic, lines/points are the greatest shapes.
But I think that's a stretch.
Calling a line a shape is a little like calling bald a hair colour. Triangles for the win.